refactored app.js and article component;
